Improve sbt build files. Fix warnings.

+++ logs_analyzer/app/scala/src/main/scala/com/databricks/apps/logs/LogAnalyzerWindowed.scala
@@ -6,7 +6,7 @@ import org.apache.spark.streaming.Seconds
 import org.apache.spark.streaming.dstream.DStream
 
 /** Collects statistics within a window. */
-class LogAnalyzerWindowed(val windowLength: Int, val slideInterval: Int) extends AnalyzeFunctions with Serializable {
+class LogAnalyzerWindowed(val windowLength: Long, val slideInterval: Long) extends AnalyzeFunctions with Serializable {
 
   import LogStatistics.EMPTY_LOG_STATISTICS
